+++ src/constants/constants.ts
@@ -16,7 +16,7 @@ export const PATH_URL = {
   MY_PORTFOLIO: '/myportfolio',
   MY_PORTFOLIO_PATH: '/myportfolio/:id',
   CREATE_PORTFOLIO: '/myportfolio/create',
-  KAKAO_AUTH: '/auth/kakao',
+  KAKAO_AUTH: '/api/users/kakao',
 };
 
 export const SERVER_URL = import.meta.env.VITE_API_URL;
